Refactor leaderboard screen to use MVP pattern
